Added OpenGl Rendering Context

+++ Lucky/Source/Lucky/ImGui/ImGuiLayer.cpp
@@ -74,7 +74,7 @@ namespace Lucky
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::GetInstance();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());	// 设置窗口大小
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());	// 设置窗口大小
 
 		// Rendering
 		ImGui::Render();
